requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.1
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.6
yarl==1.9.2
aiohttp==3.8.5
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.1.0
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.1
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tornado==6.3.3
tqdm==4.66.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.6
yarl==1.9.2
